# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateTaksPayload:
    properties:
      description:
        type: string
      projectId:
        type: string
      title:
        type: string
      userId:
        type: string
    required:
    - projectId
    - title
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateTeamRequest:
    properties:
      name:
        example: Development Team
        type: string
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.DeleteTaskResponse:
    properties:
      deleted:
        example: true
        type: boolean
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.GetTasksResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      projectId:
        type: string
      projectName:
        type: string
      status:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.Taskstatus'
      title:
        type: string
      updatedAt:
        type: string
      userId:
        $ref: '#/definitions/uuid.NullUUID'
      userName:
        type: string
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.TasksListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.GetTasksResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.Pagination'
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.TeamExistsResponse:
    properties:
      exists:
        example: true
        type: boolean
      team:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.Team'
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UpdateTaskPayload:
    properties:
      description:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.Taskstatus'
        enum:
        - ToDo
        - InProgress
        - Done
      title:
        type: string
      userId:
        type: string
    required:
    - status
    - title
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_models.Taskstatus:
    enum:
    - ToDo
    - InProgress
    - Done
    type: string
    x-enum-varnames:
    - TaskstatusToDo
    - TaskstatusInProgress
    - TaskstatusDone
  github_com_TobiasRV_challenge-fs-senior_internals_models.Team:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      ownerId:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorBody:
    properties:
      body:
        example: error message
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse:
    properties:
      errors:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorBody'
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_utils.Pagination:
    properties:
      next_cursor:
        example: eyJpZCI6IjEyMzQ1In0=
        type: string
      prev_cursor:
        example: eyJpZCI6IjEyMzQ1In0=
        type: string
    type: object
  uuid.NullUUID:
    properties:
      uuid:
        type: string
      valid:
        description: Valid is true if UUID is not NULL
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing teams, projects, tasks, and users
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Challenge FS Senior API
  version: "1.0"
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: Get paginated list of tasks. Members can only see their own tasks.
      parameters:
      - description: Filter by project ID (required for Admin/Manager)
        in: query
        name: projectId
        type: string
      - description: Filter by title
        in: query
        name: title
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Number of items per page
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.TasksListResponse'
        "400":
          description: Bad request - projectId required for Admin/Manager
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new task (Manager only)
      parameters:
      - description: Task creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateTaksPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.GetTasksResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Manager only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task by ID (Manager only)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.DeleteTaskResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Manager only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update task information (Manager only)
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UpdateTaskPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.GetTasksResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Manager only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - Tasks
  /teams:
    post:
      consumes:
      - application/json
      description: Create a new team (Admin only)
      parameters:
      - description: Team creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.Team'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "409":
          description: User already has a team
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new team
      tags:
      - Teams
  /teams/by-owner:
    get:
      consumes:
      - application/json
      description: Get the team owned by the current user (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.TeamExistsResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get team by owner
      tags:
      - Teams
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"